+++ src/config.rs
@@ -29,6 +29,10 @@ pub struct Completion {
     /// Supports `${workspaceFolder}`, `${document}`, `${userHome}` as placeholders
     #[serde(alias = "basePath")]
     pub base_path: Vec<String>,
+
+    /// Whether to automatically trigger next completion after selecting an item.
+    #[serde(alias = "triggerNextCompletion")]
+    pub trigger_next_completion: bool,
 }
 
 impl Completion {
@@ -70,11 +74,12 @@ impl Default for Config {
                 max_results: 0,
                 show_hidden_files: true,
                 exclude: vec![
-                    "**/node_modules/**".into(),
-                    "**/.git/**".into(),
+                    "**/node_modules".into(),
+                    "**/.git".into(),
                     "**/.DS_Store".into(),
                 ],
                 base_path: vec!["${workspaceFolder}".into(), "${document}".into()],
+                trigger_next_completion: true,
             },
         }
     }
@@ -96,7 +101,7 @@ pub async fn get(client: &tower_lsp::Client) -> Config {
         }])
         .await;
     let Ok(configs) = configs else {
-        info(format!(
+        warn(format!(
             "Failed to get configuration:{}, use default",
             configs.unwrap_err()
         ))
@@ -105,7 +110,7 @@ pub async fn get(client: &tower_lsp::Client) -> Config {
     };
     assert!(configs.len() == 1);
     let Ok(config) = Config::try_from(configs[0].clone()) else {
-        info(format!(
+        warn(format!(
             "Failed to parse configuration:{}, use default",
             configs[0].clone()
         ))
@@ -126,7 +131,7 @@ mod tests {
         assert!(cfg.completion.show_hidden_files);
         assert_eq!(
             cfg.completion.exclude,
-            vec!["**/node_modules/**", "**/.git/**", "**/.DS_Store"]
+            vec!["**/node_modules", "**/.git", "**/.DS_Store"]
         );
         assert_eq!(
             cfg.completion.base_path,
@@ -145,6 +150,7 @@ mod tests {
                 "${document}".into(),
                 "/absolute/path".into(),
             ],
+            trigger_next_completion: true,
         };
 
         let workspace_folders = vec!["/ws1".to_string(), "/ws2".to_string()];
@@ -170,6 +176,7 @@ mod tests {
             show_hidden_files: true,
             exclude: vec![],
             base_path: vec!["${document}".into(), "${userHome}/foo".into()],
+            trigger_next_completion: true,
         };
 
         let workspace_folders = vec![];