+++ src/main.rs
@@ -2,6 +2,10 @@ use path_server::PathServer;
 
 #[tokio::main]
 async fn main() {
+    unsafe {
+        std::env::set_var("RUST_BACKTRACE", "1");
+    }
+
     let stdin = tokio::io::stdin();
     let stdout = tokio::io::stdout();